request-new water-api/requirements.txt
alembic==1.9.2
apispec==6.0.2
charset-normalizer==3.0.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
Flask-Migrate==4.0.1
flask-smorest==0.40.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mysql-client==0.0.1
num2words==0.5.12
packaging==23.0
PyMySQL==1.0.2
pyotp==2.8.0
python-dotenv==0.21.0
requests==2.28.2
SQLAlchemy==1.4.46
urllib3==1.26.14
webargs==8.2.0
Werkzeug==2.2.2